+++ endpoints/payments/delete.php
@@ -1,13 +1,14 @@
 <?php
 
 require_once '../../includes/connect_endpoint.php';
-session_start();
+
 if (isset($_SESSION['loggedin']) && $_SESSION['loggedin'] === true) {
     if ($_SERVER["REQUEST_METHOD"] === "DELETE") {
         $paymentMethodId = $_GET["id"];
-        $deleteQuery = "DELETE FROM payment_methods WHERE id = :paymentMethodId";
+        $deleteQuery = "DELETE FROM payment_methods WHERE id = :paymentMethodId and user_id = :userId";
         $deleteStmt = $db->prepare($deleteQuery);
         $deleteStmt->bindParam(':paymentMethodId', $paymentMethodId, SQLITE3_INTEGER);
+        $deleteStmt->bindParam(':userId', $userId, SQLITE3_INTEGER);
 
         if ($deleteStmt->execute()) {
             $success['success'] = true;
