+++ endpoints/settings/resettheme.php
@@ -1,7 +1,7 @@
 <?php
 
     require_once '../../includes/connect_endpoint.php';
-    session_start();
+    
     if (!isset($_SESSION['loggedin']) || $_SESSION['loggedin'] !== true) {
         die(json_encode([
             "success" => false,
@@ -10,7 +10,8 @@
     }
 
     if ($_SERVER["REQUEST_METHOD"] === "DELETE") {
-        $stmt = $db->prepare('DELETE FROM custom_colors');
+        $stmt = $db->prepare('DELETE FROM custom_colors WHERE user_id = :userId');
+        $stmt->bindParam(':userId', $userId, SQLITE3_INTEGER);
 
         if ($stmt->execute()) {
             die(json_encode([
